requirements.txt
glfw==2.5.0
gym==0.19.0
h5py==3.6.0
imageio==2.14.1
matplotlib==3.5.1
numpy==1.21.5
opencv-python==4.5.5.62
Pillow==9.0.0
pybind11==2.9.0
pygifsicle==1.0.5
PyOpenGL==3.1.5
PyOpenGL-accelerate==3.1.5
stable-baselines3==1.4.0
#--find-links https://download.pytorch.org/whl/torch_stable.html
#torch==1.11.0+cu102 
#torchvision==0.12.0+cu102
torch==1.11.0
torchvision==0.12.0 
torchaudio==0.11.0 
ttkbootstrap==1.5.1
typing==3.7.4.3
seaborn==0.9.0
tensorboard